+++ src/main/java/ru/mts/loanservice/exception/TariffNotFound.java
@@ -1,6 +1,6 @@
 package ru.mts.loanservice.exception;
 
-public class TariffNotFound extends RuntimeException{
+public class TariffNotFound extends RuntimeException {
 
     public TariffNotFound(String message) {
         super(message);